Windows does not support Unix sockets
Add conditional compilation to support Unix sockets on non Windows
platforms only.
