+++ api/server.go
@@ -10,10 +10,8 @@ import (
 	"io"
 	"net"
 	"net/http"
-	"os"
 	"strconv"
 	"strings"
-	"syscall"
 	"time"
 
 	log "github.com/Sirupsen/logrus"
@@ -161,29 +159,6 @@ func NewServer(proto, addr string) (*Server, error) {
 func setupTCPHTTP(addr string) (*Server, error) {
 	l, err := net.Listen("tcp", addr)
 	if err != nil {
-		return nil, err
-	}
-
-	a := &Server{
-		HTTPServer: &http.Server{Addr: addr},
-		Listener:   l,
-	}
-	return a, nil
-}
-
-func setupUnixHTTP(addr string) (*Server, error) {
-	if err := syscall.Unlink(addr); err != nil && !os.IsNotExist(err) {
-		return nil, err
-	}
-	mask := syscall.Umask(0777)
-	defer syscall.Umask(mask)
-
-	l, err := net.Listen("unix", addr)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := os.Chmod(addr, 0660); err != nil {
 		return nil, err
 	}
